agregue cartera, colegiatura, expenses, pagos, y personal
